requirements.txt
py==1.4.31
pytest==2.9.2
PyYAML==3.11
